Revert "74 refresh database"
