+++ src/dashowl/issues/views.py
@@ -4,8 +4,6 @@ import requests
 import json
 from .models import Issue
 from .. import secret
-from ..repositories.views import REPO_ATUAL
-from ..repositories.models import Repository
 
 
 #É nescessario passar para as todas essas funcoes o request ?
@@ -14,20 +12,16 @@ from ..repositories.models import Repository
 
 def get_issues(request):
 
-    user_login = request.session['login']
-    token = request.session['token']
-
-    # g = Github(login_or_token=token)
-    g = Github(login_or_token=user_login, password=secret.password)
-    repo = g.get_repo(full_name_or_id=REPO_ATUAL)
-
+    g = Github(secret.login, secret.password)
+    repo = g.get_repo("fga-eps-mds/2019.2-DashboardAgil-Wiki")
     """
         dar um jeito de pegar o id do repositório atual
     """
-    repository = Repository.objects.get(repositoryID=repo.id)
+
+
     if bool(Issue.objects.filter(repository__repositoryID=repo.id)):
-        all_issues = Issue.objects.filter(repository__repositoryID=repo.id).order_by('date')
-        refresh_issues(repo, repository, list(all_issues)[-1].date)
+        all_issues = Issue.objects.filter(repository__repositoryID=repo.id)
+
         authores = all_issues.distinct('author')
 
         a = []
@@ -74,10 +68,7 @@ def get_issues(request):
 
 
     else:
-        save_issue(repo, repository)
-        open_issues = Issue.objects.filter(repository__repositoryID=repo.id, state='open')
-        closed_issues = Issue.objects.filter(repository__repositoryID=repo.id, state='closed')
-        all_issues = Issue.objects.filter(repository__repositoryID=repo.id)
+        raise TypeError
 
 #Retorna o valor de cada issue
     req= requests.get('https://api.zenhub.io/p1/repositories/206358281/issues/39?access_token=02a009e06e4926091eadce6ef1dffc9f9b3f7b5bd417b116ea90c55bf6fb68dda7eb367ab6544c07')
@@ -92,23 +83,3 @@ def get_issues(request):
     'date11c': date11c,'date12c': date12c})
 
 
-def refresh_issues(repo, repository, last):
-    for issue in repo.get_issues(state='all', since=last):
-        issues_model = Issue.objects.create(repository=repository,
-                                            issue_number=issue.number,
-                                            state=issue.state,
-                                            author=issue.user.login,
-                                            date=issue.created_at)
-        issues_model.publish()
-
-
-def save_issue(repo, repository):
-    issues = repo.get_issues(state='all')
-    if bool(list(issues)):
-        for issue in issues:
-            issues_model = Issue.objects.create(repository=repository,
-                                                issue_number=issue.number,
-                                                state=issue.state,
-                                                author=issue.user.login,
-                                                date=issue.created_at)
-            issues_model.publish()
