+++ src/dashowl/repositories/views.py
@@ -9,8 +9,6 @@ from ..issues.models import Issue
 from ..pull_requests.models import Pull_request
 
 
-REPO_ATUAL = "fga-eps-mds/2019.2-DashboardAgil"
-
 def repositories(request):
 
     user_login = request.session['login']
@@ -19,9 +17,6 @@ def repositories(request):
     repos_names = []
     repos_id = []
     if bool(Repository.objects.filter(user__login=user_login)):
-        user = Usuario.objects.get(login=user_login)
-        repos = Repository.objects.filter(user__login=user_login)
-        refresh_repos(g, user, list(repos)[-1].id)
         repos = Repository.objects.filter(user__login=user_login)
         for repository in repos:
             repos_names.append(repository.name)
@@ -35,6 +30,10 @@ def repositories(request):
             repository_model.publish()
             repos_names.append(repository.name)
             repos_id.append(repository.id)
+            save_commit(repository, repository_model)
+            save_issue(repository, repository_model)
+            save_milestone(repository, repository_model)
+            save_pull_request(repository, repository_model)
         
         for repository in g.get_user().get_repos(type='member'):
             repository_model = Repository.objects.create(user=user,
@@ -43,7 +42,11 @@ def repositories(request):
             repository_model.publish()
             repos_names.append(repository.name)
             repos_id.append(repository.id)
-  
+            save_commit(repository, repository_model)
+            save_issue(repository, repository_model)
+            save_milestone(repository, repository_model)
+            save_pull_request(repository, repository_model)
+
         for repository in g.get_user().get_repos(type='private'):
             repository_model = Repository.objects.create(user=user,
                                                          name=repository.name,
@@ -51,16 +54,52 @@ def repositories(request):
             repository_model.publish()
             repos_names.append(repository.name)
             repos_id.append(repository.id)
+            save_commit(repository, repository_model)
+            save_issue(repository, repository_model)
+            save_milestone(repository, repository_model)
+            save_pull_request(repository, repository_model)
 
     repos = zip(repos_names, repos_id)
 
     return render(request, 'repositories.html', {'repos': repos})
 
 
-def refresh_repos(g, user, last):
-    repos = g.get_user().get_repos(type='member')
-    for i in range(last+1, repos.totalCount):
-        repository_model = Repository.objects.create(user=user,
-                                                     name=repos[i].name,
-                                                     repositoryID=repos[i].id)
-        repository_model.publish()
+def save_commit(repo, repository):
+    for commit in repo.get_commits():
+        commit_model = Commit.objects.create(repository=repository,
+                                             sha_commit=commit.sha,
+                                             author=commit.commit.author.name,
+                                             date=commit.commit.author.date)
+        commit_model.publish()
+
+
+def save_issue(repo, repository):
+    for issue in repo.get_issues(state='all'):
+        issues_model = Issue.objects.create(repository=repository,
+                                            issue_number=issue.number,
+                                            state=issue.state,
+                                            author=issue.user.login,
+                                            date=issue.created_at)
+        issues_model.publish()
+
+
+def save_milestone(repo, repository):
+    for milestone in repo.get_milestones(state='all'):
+        milestone_model = Milestone.objects.create(repository=repository,
+                                                   milestoneID=milestone.id,
+                                                   state=milestone.state,
+                                                   title=milestone.title,
+                                                   author=milestone.creator.login,
+                                                   created_at=milestone.created_at,
+                                                   due_on=milestone.due_on)
+        milestone_model.publish()
+
+
+def save_pull_request(repo, repository):
+    for pull_request in repo.get_pulls(state='all'):
+        pull_requests_model = Pull_request.objects.create(repository=repository,
+                                                          pull_request_number=pull_request.number,
+                                                          state=pull_request.state,
+                                                          author=pull_request.user.login,
+                                                          open_date=pull_request.created_at)
+        pull_requests_model.publish()
